requirements.txt
fastapi~=0.115.8
pydantic~=2.10.6
mysql-connector-python~=9.2.0
streamlit~=1.41.1
requests~=2.32.3
pandas~=2.2.3
click~=8.1.8